// js/app/src/store/__generated__/datasetStore_latestVersionQuery.graphql.ts
export type datasetStore_latestVersionQuery$variables = {
  datasetId: string;
};
export type datasetStore_latestVersionQuery$data = {
  readonly dataset: {
    readonly id: string;
    readonly latestVersions?: {
      readonly edges: ReadonlyArray<{
        readonly version: {
          readonly createdAt: string;
          readonly description: string | null;
          readonly id: string;
        };
      }>;
    };
  };
};
export type datasetStore_latestVersionQuery = {
  response: datasetStore_latestVersionQuery$data;
  variables: datasetStore_latestVersionQuery$variables;
};